Add Hosts override
