+++ lib/l10n/intl/messages_zh_CN.dart
@@ -45,6 +45,7 @@ class MessageLookup extends MessageLookupByLibrary {
         "allowLanDesc": MessageLookupByLibrary.simpleMessage("允许通过局域网访问代理"),
         "app": MessageLookupByLibrary.simpleMessage("应用"),
         "appAccessControl": MessageLookupByLibrary.simpleMessage("应用访问控制"),
+        "appDesc": MessageLookupByLibrary.simpleMessage("处理应用相关设置"),
         "application": MessageLookupByLibrary.simpleMessage("应用程序"),
         "applicationDesc": MessageLookupByLibrary.simpleMessage("修改应用程序相关设置"),
         "auto": MessageLookupByLibrary.simpleMessage("自动"),
@@ -105,6 +106,7 @@ class MessageLookup extends MessageLookupByLibrary {
         "direct": MessageLookupByLibrary.simpleMessage("直连"),
         "discoverNewVersion": MessageLookupByLibrary.simpleMessage("发现新版本"),
         "discovery": MessageLookupByLibrary.simpleMessage("发现新版本"),
+        "dnsDesc": MessageLookupByLibrary.simpleMessage("更新DNS相关设置"),
         "doYouWantToPass": MessageLookupByLibrary.simpleMessage("是否要通过"),
         "download": MessageLookupByLibrary.simpleMessage("下载"),
         "edit": MessageLookupByLibrary.simpleMessage("编辑"),
@@ -129,6 +131,7 @@ class MessageLookup extends MessageLookupByLibrary {
             MessageLookupByLibrary.simpleMessage("开启后存在闪退风险"),
         "fourColumns": MessageLookupByLibrary.simpleMessage("四列"),
         "general": MessageLookupByLibrary.simpleMessage("基础"),
+        "generalDesc": MessageLookupByLibrary.simpleMessage("覆写基础设置"),
         "geoData": MessageLookupByLibrary.simpleMessage("地理数据"),
         "geodataLoader": MessageLookupByLibrary.simpleMessage("Geo低内存模式"),
         "geodataLoaderDesc":
@@ -146,6 +149,8 @@ class MessageLookup extends MessageLookupByLibrary {
         "just": MessageLookupByLibrary.simpleMessage("刚刚"),
         "keepAliveIntervalDesc":
             MessageLookupByLibrary.simpleMessage("TCP保持活动间隔"),
+        "key": MessageLookupByLibrary.simpleMessage("键"),
+        "keyNotEmpty": MessageLookupByLibrary.simpleMessage("键不能为空"),
         "language": MessageLookupByLibrary.simpleMessage("语言"),
         "layout": MessageLookupByLibrary.simpleMessage("布局"),
         "light": MessageLookupByLibrary.simpleMessage("浅色"),
@@ -292,7 +297,10 @@ class MessageLookup extends MessageLookupByLibrary {
         "upload": MessageLookupByLibrary.simpleMessage("上传"),
         "url": MessageLookupByLibrary.simpleMessage("URL"),
         "urlDesc": MessageLookupByLibrary.simpleMessage("通过URL获取配置文件"),
+        "value": MessageLookupByLibrary.simpleMessage("值"),
+        "valueNotEmpty": MessageLookupByLibrary.simpleMessage("值不能为空"),
         "view": MessageLookupByLibrary.simpleMessage("查看"),
+        "vpnDesc": MessageLookupByLibrary.simpleMessage("修改VPN相关设置"),
         "webDAVConfiguration": MessageLookupByLibrary.simpleMessage("WebDAV配置"),
         "whitelistMode": MessageLookupByLibrary.simpleMessage("白名单模式"),
         "years": MessageLookupByLibrary.simpleMessage("年"),
